Gameplay Update 1.15
-Added Chronosphere combo
	-3x Stun spells
	-Creates a 3x3 bubble in time 4 squares ahead, stunning them, and
making them take double damage
	-It's also faceless void's ult
-Added Debuff Transfer spell
	-Removes your debuffs and transfers them to all enemy units it hits
	-Each debuff purged makes the projectile stronger
-Added Sanctuary spell
	-Makes all the panels on your side Holy panels, halving damage and
doubling healing
-Turned Regenerate into a status effect (that can be purged)
-AWP shot can now penetrate shields
-Damage Amplification damage multiplier from 2 to 1.5
-Damage Amplification duration from 10 to 15
-Electro Bolt now purges buffs as well
-Wish Upon a Lucky Star now penetrates and obliterates all shields

+++ SpellWars/src/projectile/CrackGrenade.java
@@ -14,7 +14,7 @@ import unit.Unit;
 public class CrackGrenade extends Grenade {
 	float crackChance;
 	Trail trail;
-	public CrackGrenade(int damage, float crackChance, double duration, Point endDisplacement, float initialHeight, float endHeight, Point gridLoc, String imagePath, int teamID) {
+	public CrackGrenade(double damage, float crackChance, double duration, Point endDisplacement, float initialHeight, float endHeight, Point gridLoc, String imagePath, int teamID) {
 		super(damage, duration, endDisplacement, initialHeight, endHeight, gridLoc, imagePath, teamID);
 		this.crackChance = crackChance;
 	}
